Merged PR 13474: Unificación de provider

# [NGM-6813](https://acelerabga.atlassian.net/browse/NGM-6813) - [BFF] Unificación de AccountStatementProvider y OwnAccountMiddlewareProvider

## Cambios

- new AccountStatementsMiddlewareProvider
- Services
- Controller
- Mapper
- Test

Cherry-picked from commit `82bf49a2`.

+++ bff/src/main/java/bg/com/bo/bff/application/dtos/request/account/statement/AmountRange.java
@@ -6,14 +6,16 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class AmountRange {
     @Schema(example = "0.5", description = "Monto mínimo")
-    private Double min;
+    private BigDecimal min;
 
     @Schema(example = "1000", description = "Monto máximo")
-    private Double max;
+    private BigDecimal max;
 }
